requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
apscheduler>=3.10.0
requests>=2.31.0
python-dotenv==1.0.1
duckduckgo-search==6.3.1
beautifulsoup4==4.12.3
aiohttp==3.9.5
fake-useragent==1.5.1
tenacity==9.0.0
lxml==5.2.1
